Split rust source code into multiple files.
