+++ zowex/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "zowe"
-version = "0.9.0"
+version = "0.9.1"
 authors = ["Zowe Project"]
 edition = "2018"
 license = "EPL-2.0"
